backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.0
django-environ==0.11.2
mysqlclient==2.2.0
gunicorn==21.2.0
whitenoise==6.6.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
torch==2.1.0
torchvision==0.16.0
numpy==1.24.3
